+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `supported_extensions` to Vendor, allowing parsers to specify supported file extensions.
 - Added support for `multiple read modes` in `Agilent Gen5` Adapter
+- Added error message for zero plate reader documents in `Softmax Pro` Adapter
 
 ### Fixed